修改倒计时问题，修改订单文字反馈，添加涨跌幅

+++ app/src/main/java/com/yundian/star/been/TimeLineBeen.java
@@ -33,7 +33,7 @@ public class TimeLineBeen {
         private int closedYesterdayPrice;
         private double currentPrice;
         private int openingTodayPrice;
-        private int pchg;
+        private float pchg;
         private int priceTime;
         private String symbol;
 
@@ -69,11 +69,11 @@ public class TimeLineBeen {
             this.openingTodayPrice = openingTodayPrice;
         }
 
-        public int getPchg() {
+        public float getPchg() {
             return pchg;
         }
 
-        public void setPchg(int pchg) {
+        public void setPchg(float pchg) {
             this.pchg = pchg;
         }
